rework to split graph, pathfinding and game logic
