+++ src/utils.ts
@@ -1,0 +1,4 @@
+export interface Coordinates {
+    x: number;
+    y: number;
+}
